using builtin picker for account picker view and refactored code
